+++ src/cmd_spanr/combine.rs
@@ -18,6 +18,14 @@ pub fn make_subcommand<'a, 'b>() -> App<'a, 'b> {
                 .index(1),
         )
         .arg(
+            Arg::with_name("op")
+                .long("op")
+                .takes_value(true)
+                .default_value("union")
+                .empty_values(false)
+                .help("Operations: intersect, union, diff or xor"),
+        )
+        .arg(
             Arg::with_name("outfile")
                 .short("o")
                 .long("outfile")
@@ -37,19 +45,35 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), std::io::Error> {
     let s_of = yaml2set_m(&yaml);
     let chrs = chrs_in_sets(&s_of);
 
+    let op = args.value_of("op").unwrap();
+
     //----------------------------
     // Operating
     //----------------------------
     let mut res: BTreeMap<String, IntSpan> = BTreeMap::new();
-    for chr in &chrs {
-        res.insert(chr.to_string(), IntSpan::new());
-    }
+    fill_up_s(&mut res, &chrs);
 
-    for name in s_of.keys() {
+    let names: Vec<_> = s_of.keys().cloned().collect();
+    let first = names[0].clone();
+
+    for name in names {
         let set = s_of.get(name.as_str()).unwrap();
         for chr in set.keys() {
-            let intspan = set.get(chr).unwrap();
-            res.entry(chr.to_string()).and_modify(|e| e.merge(intspan));
+            if name == first {
+                let intspan = set.get(chr).unwrap();
+                res.entry(chr.to_string()).and_modify(|e| e.merge(intspan));
+            } else {
+                let mut intspan_op = res.get(chr).unwrap().copy();
+                intspan_op = match op {
+                    "intersect" => intspan_op.intersect(set.get(chr).unwrap()),
+                    "diff" => intspan_op.diff(set.get(chr).unwrap()),
+                    "union" => intspan_op.union(set.get(chr).unwrap()),
+                    "xor" => intspan_op.xor(set.get(chr).unwrap()),
+                    _ => panic!("Invalid IntSpan Op"),
+                };
+                //                eprintln!("Op {}: {}", op, intspan_op.to_string());
+                res.insert(chr.into(), intspan_op);
+            }
         }
     }